flask-app/requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
PyMySQL==1.1.0

# Authentication
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Authlib==1.3.0

# Forms and Validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
marshmallow==3.20.1

# Email
Flask-Mail==0.9.1

# Admin Panel
Flask-Admin==1.6.1

# API
Flask-RESTX==1.3.0
Flask-CORS==4.0.0

# Payment Gateways
paypalrestsdk==1.13.1

# Async Tasks
celery==5.3.4
redis==5.0.1
Flask-Caching==2.1.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0

# Security
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0

# Production Server
gunicorn==21.2.0
